% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_bb.R
\name{bb_post}
\alias{bb_post}
\title{Posterior inference in beta-binomial model.}
\usage{
bb_post(ncounts, ssize, prior, seq_error = 0.01, log = FALSE)
}
\arguments{
\item{ncounts}{A vector of non-negative integers. The ith element
is the number of counts of the ith sample.}

\item{ssize}{A vector of positive integers. The ith element is the
total number of counts of the ith sample.}

\item{prior}{A vector of non-negative numerics that sum to one. The
prior probability on the genotype. The first element is the
prior probability of zero reference alleles, the second element
is the prior probability of one reference allele, etc. The
length of \code{prior} is one more than the ploidy of the
species. You can alternatively specify \code{prior} as the
ploidy of the individual, for which it will set a uniform prior
on the genotype. For example, setting \code{prior = 3} will
result in using \code{(1/4, 1/4, 1/4, 1/4)} as the prior
probability for the genotypes (Aaaa, AAaa, AAAa, AAAA) where
"A" is the reference allele in a 4-ploid individual.}

\item{seq_error}{A non-negative numeric. This is the known
sequencing error rate. This is a rough high-ball error rate
given by Li et. al. (2011).}

\item{log}{A logical. Should we return the log probabilties (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
The main difference between this function and \code{bin_post} is that here we also estimate an
overdispersion parameter by maximum marginal likelihood.
}
\author{
David Gerard
}

