% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{bb_simple_post}
\alias{bb_simple_post}
\title{Simple posterior inference for beta-binomial when ncounts and ssize are only length 1.}
\usage{
bb_simple_post(ncounts, ssize, ploidy, p1geno, p2geno, seq_error = 0,
  bias_val = 1, od_param = 0)
}
\arguments{
\item{ncounts}{A vector of non-negative integers. The ith element
is the number of counts of the ith sample.}

\item{ssize}{A vector of positive integers. The ith element is the
total number of counts of the ith sample.}

\item{ploidy}{The ploidy of the species.}

\item{p1geno}{The first parental genotype.}

\item{p2geno}{The second parental genotype.}

\item{seq_error}{A non-negative numeric. This is the known
sequencing error rate. This is a rough high-ball error rate
given by Li et. al. (2011).}

\item{bias_val}{The bias parameter. A value of 1 means no bias.}

\item{od_param}{The over-dispersion parameter. A value of 0 means no overdispersion.
A value of 1 means total overdispersion.}
}
\description{
Simple posterior inference for beta-binomial when ncounts and ssize are only length 1.
}
\author{
David Gerard
}
