% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbpost_double}
\alias{bbpost_double}
\title{Posterior inference under the beta-binomial model with outliers. If \code{od_param = 0}, then
this reduces to the binomial model with outliers.}
\usage{
bbpost_double(x, n, ploidy, p1geno, p2geno, seq_error, bias_val, od_param,
  outlier = FALSE, out_prop = 0.1, out_mean = 0.5, out_disp = 1/3)
}
\arguments{
\item{x}{The observed number of reference alleles.}

\item{n}{The total number of reads.}

\item{ploidy}{The ploidy of the species}

\item{p1geno}{The first parental genotype (number of reference alleles parent 1 has).}

\item{p2geno}{The second parental genotype (number of reference alleles parent 2 has).}

\item{seq_error}{The sequencing error rate.}

\item{bias_val}{The bias parameter. A value of 1 indicates no bias.}

\item{od_param}{The over-dispersion parameter. Should be in [0, 1). A value of 0 reduces to the
binomial model.}

\item{outlier}{A logical. Should we include the outlier model (\code{TRUE}) or not (\code{FALSE})?}

\item{out_prop}{The proportion of observations that are outliers.}

\item{out_mean}{The mean of outlier distribution.}

\item{out_disp}{The overdispersion parameter of the outlier distribution.}
}
\description{
Posterior inference under the beta-binomial model with outliers. If \code{od_param = 0}, then
this reduces to the binomial model with outliers.
}
\author{
David Gerard
}
