% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbpost_tot}
\alias{bbpost_tot}
\title{Posterior inference for each individual.}
\usage{
bbpost_tot(ocounts, osize, ploidy, prob_geno, seq_error, bias_val, od_param,
  outlier = FALSE, out_prop = 0.1, out_mean = 0.5, out_disp = 1/3)
}
\arguments{
\item{ocounts}{The vector of counts of the reference allele of the offspring.}

\item{osize}{The vector of total reads of the offspring.}

\item{ploidy}{The ploidy of the species}

\item{prob_geno}{The distribution of genotypes.}

\item{seq_error}{The sequencing error rate.}

\item{bias_val}{The bias parameter. A value of 1 indicates no bias.}

\item{od_param}{The over-dispersion parameter. Should be in [0, 1). A value of 0 reduces to the
binomial model.}

\item{outlier}{A logical. Should we include the outlier model (\code{TRUE}) or not (\code{FALSE})?}

\item{out_prop}{The proportion of observations that are outliers.}

\item{out_mean}{The mean of outlier distribution.}

\item{out_disp}{The overdispersion parameter of the outlier distribution.}
}
\description{
Posterior inference for each individual.
}
\author{
David Gerard
}
