% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_algorithm.R
\name{bin_post}
\alias{bin_post}
\title{Calculates posterior probabilities of a genotype given just the sequence
counts.}
\usage{
bin_post(ncounts, ssize, prior)
}
\arguments{
\item{ncounts}{A vector of non-negative integers. The ith element is the
number of counts of the ith sample.}

\item{ssize}{A vector of positive integers. The ith element is the total
number of counts of the ith sample.}

\item{prior}{A vector of non-negative numerics that sum to one. The prior
probability on the genotype. The first element is the prior probability of
zero reference alleles, the second element is the prior probability of one
reference allele, etc. The length of \code{prior} is one more than the
ploidy of the species. You can alternatively specify \code{prior} as the
ploidy of the individual, for which it will set a uniform prior on the
genotype. For example, setting \code{prior = 3} will result in using
\code{(1/4, 1/4, 1/4, 1/4)} as the prior probability for the genotypes
(Aaaa, AAaa, AAAa, AAAA) where "A" is the reference allele in a 4-ploid
individual.}
}
\description{
The function assumes that all data come from independent samples on the same
individual.
}
\author{
David Gerard
}

