% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colSums_cpp}
\alias{colSums_cpp}
\title{Stupid implementation of colSums because I guess not implemented in Rcpp sugar.}
\usage{
colSums_cpp(x)
}
\arguments{
\item{x}{A NumericMatrix.}
}
\description{
Stupid implementation of colSums because I guess not implemented in Rcpp sugar.
}
\author{
David Gerard
}
