% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_dd}
\alias{dbeta_dd}
\title{Derivative of betabinomial density w.r.t. original bias parameter.}
\usage{
dbeta_dd(x, n, d, eps, p, tau)
}
\arguments{
\item{x}{The number of counts of reference allele.}

\item{n}{The number of counts of reads.}

\item{d}{The sequencing bias parameter.}

\item{eps}{The sequencing error rate}

\item{p}{The proportion of genome that is the reference allele.}

\item{tau}{The overdisperion parameter.}
}
\description{
Uses chain rule with \code{\link{dbeta_dprop}} * \code{\link{dxi_dd}}.
}
\author{
David Gerard
}
