% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_deps}
\alias{dbeta_deps}
\title{Derivative of beta density w.r.t. sequencing error rate.}
\usage{
dbeta_deps(x, n, d, eps, p, tau)
}
\arguments{
\item{x}{The number of counts of reference allele.}

\item{n}{The number of counts of reads.}

\item{d}{The sequencing bias parameter.}

\item{eps}{The sequencing error rate}

\item{p}{The proportion of genome that is the reference allele.}

\item{tau}{The overdisperion parameter.}
}
\description{
I use the chain rule here. \code{\link{dbeta_dprop}} * \code{\link{dxi_df}} *
\code{\link{df_deps}}.
}
\author{
David Gerard
}
