% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_dh}
\alias{dbeta_dh}
\title{Derivative of overdispersion parameter of beta density.}
\usage{
dbeta_dh(x, n, xi, h)
}
\arguments{
\item{x}{The observed counts of reference allele.}

\item{n}{The total number of observed counts.}

\item{xi}{The mean proportion.}

\item{h}{A parameterization of the overdispersion parameter.
This is just constrained to be greater than 0.}
}
\description{
h = (1 - tau) / tau, so tau = 1 / (h + 1).
}
