% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_dl}
\alias{dbeta_dl}
\title{Derivative of beta density w.r.t. unconstrained parameterization of sequencing error.}
\usage{
dbeta_dl(x, n, d, ell, p, h)
}
\arguments{
\item{x}{The number of counts of reference allele.}

\item{n}{The number of counts of reads.}

\item{d}{The sequencing bias parameter.}

\item{ell}{Expit of eps. We have eps = exp(ell) / (1 + exp(ell)), or ell = log(eps / (1 - eps))}

\item{p}{The proportion of genome that is the reference allele.}

\item{h}{the overdisperion parameter. tau = 1 / (h + 1) or h = (1 - tau) / tau}
}
\description{
I use the chain rule here. \code{\link{dbeta_dprop}} * \code{\link{dxi_df}} *
\code{\link{df_deps}} * \code{\link{deps_dell}}.
}
\author{
David Gerard
}
