% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_dprop}
\alias{dbeta_dprop}
\title{Derivative of mean of beta density}
\usage{
dbeta_dprop(x, n, xi, tau)
}
\arguments{
\item{x}{The observed counts of reference allele.}

\item{n}{The total number of observed counts.}

\item{xi}{The mean proportion.}

\item{tau}{The overdispersion parameter.}
}
\description{
Derivative of mean of beta density
}
\author{
David Gerard
}
