% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_dr}
\alias{dbeta_dr}
\title{Derivative of beta(x|n, xi, r), where r = log(h) from \code{\link{dbeta_dh}}.}
\usage{
dbeta_dr(x, n, xi, r)
}
\arguments{
\item{x}{The observed counts of reference allele.}

\item{n}{The total number of observed counts.}

\item{xi}{The mean proportion.}

\item{r}{We have \code{r = log(h)} from double dbeta_dr().}
}
\description{
Derivative of beta(x|n, xi, r), where r = log(h) from \code{\link{dbeta_dh}}.
}
\author{
David Gerard
}
