% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_dr_ell}
\alias{dbeta_dr_ell}
\title{Same as \code{\link{dbeta_dh}}, but with same inputs as \code{\link{dbeta_ds}}
and \code{\link{dbeta_dl}}.}
\usage{
dbeta_dr_ell(x, n, d, ell, p, r)
}
\arguments{
\item{x}{The number of counts of reference allele.}

\item{n}{The number of counts of reads.}

\item{d}{The sequencing bias parameter.}

\item{ell}{Expit of eps. We have eps = exp(ell) / (1 + exp(ell)), or ell = log(eps / (1 - eps))}

\item{p}{The proportion of genome that is the reference allele.}

\item{r}{We have \code{tau = 1 / (1 + exp(r))}}
}
\description{
Same as \code{\link{dbeta_dh}}, but with same inputs as \code{\link{dbeta_ds}}
and \code{\link{dbeta_dl}}.
}
\author{
David Gerard
}
