% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_ds}
\alias{dbeta_ds}
\title{Derivative of betabinomial density w.r.t. bias parameter.}
\usage{
dbeta_ds(x, n, s, ell, p, h)
}
\arguments{
\item{x}{The number of counts of reference allele.}

\item{n}{The number of counts of reads.}

\item{s}{We have \code{s = exp(d)}, where \code{d} is the bias parameter.}

\item{ell}{Expit of eps. We have eps = exp(ell) / (1 + exp(ell)), or ell = log(eps / (1 - eps))}

\item{p}{The proportion of genome that is the reference allele.}

\item{h}{the overdisperion parameter. tau = 1 / (h + 1) or h = (1 - tau) / tau}
}
\description{
Uses chain rule with \code{\link{dbeta_dprop}} * \code{\link{dxi_dd}} * \code{\link{dd_ds}}.
}
\author{
David Gerard
}
