% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_dtau}
\alias{dbeta_dtau}
\title{Derivative of beta(x|n, xi, tau).}
\usage{
dbeta_dtau(x, n, d, eps, p, tau)
}
\arguments{
\item{x}{The number of counts of reference allele.}

\item{n}{The number of counts of reads.}

\item{d}{The sequencing bias parameter.}

\item{eps}{The sequencing error rate}

\item{p}{The proportion of genome that is the reference allele.}

\item{tau}{The overdisperion parameter.}
}
\description{
Derivative of beta(x|n, xi, tau).
}
\author{
David Gerard
}
