% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_bb.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Beta-binomial density function.}
\usage{
dbetabinom(x, size, alpha, beta, log = FALSE)
}
\arguments{
\item{x}{The observed counts.}

\item{size}{The number of trials.}

\item{alpha}{The alpha parameter of the underlying beta.}

\item{beta}{The beta parameter of the underlying beta}

\item{log}{A logical. Should we return the log of the density
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
Note that it is possible here for \code{alpha = 0} or for \code{beta = 0} (but not both).
In which case, this function becomes the indicator function for \code{x == 0}
(when \code{alpha = 0}) or \code{x == size} (when \code{beta = 0}).
}
\author{
David Gerard
}
