% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbetabinom_cpp}
\alias{dbetabinom_cpp}
\title{An Rcpp version of the \code{\link{dbetabinom}}.}
\usage{
dbetabinom_cpp(x, size, alpha_shape, beta_shape, return_log = FALSE)
}
\arguments{
\item{x}{A numeric vector of the observed counts.}

\item{size}{A numeric vector of the number of trials.}

\item{alpha_shape}{A numeric scalar of the alpha parameter underlying beta.}

\item{beta_shape}{A numeric scalar of the beta parameter of the underlying beta.}

\item{return_log}{A logical scalar. Should we return the log of the density (\code{TRUE})
or not (\code{FALSE})?}
}
\description{
An Rcpp version of the \code{\link{dbetabinom}}.
}
\author{
David Gerard
}
