% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_bb.R
\name{dbetabinom_mu_rho}
\alias{dbetabinom_mu_rho}
\title{beta-binomial density function in terms of mean and correlation.}
\usage{
dbetabinom_mu_rho(x, size, mu, rho, log = FALSE)
}
\arguments{
\item{x}{The observed counts.}

\item{size}{The number of trials.}

\item{mu}{The mean proportion. Equivalent to \code{alpha / (alpha + beta)} in \code{\link{dbetabinom}}.}

\item{rho}{The correlation/overdispersion parameter.
Equivalent to \code{1 / (1 + alpha + beta)} in \code{\link{dbetabinom}}.}

\item{log}{A logical. Should we return the log of the density
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
beta-binomial density function in terms of mean and correlation.
}
