% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbetabinom_mu_rho_cpp}
\alias{dbetabinom_mu_rho_cpp}
\title{An Rcpp version of \code{\link{dbetabinom_mu_rho}}.}
\usage{
dbetabinom_mu_rho_cpp(x, size, mu, rho, return_log = FALSE)
}
\arguments{
\item{x}{A numeric vector of the observed counts.}

\item{size}{A numeric vector of the number of trials.}

\item{mu}{The mean of the underlying beta.}

\item{rho}{The overdispersion parameter of the underlying beta.}

\item{return_log}{A logical scalar. Should we return the log of the density (\code{TRUE})
or not (\code{FALSE})?}
}
\description{
An Rcpp version of \code{\link{dbetabinom_mu_rho}}.
}
\author{
David Gerard
}
