% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbetabinom_mu_rho_cpp_double}
\alias{dbetabinom_mu_rho_cpp_double}
\title{double version of betabinomial density.}
\usage{
dbetabinom_mu_rho_cpp_double(x, size, mu, rho, return_log)
}
\arguments{
\item{x}{The number of successes}

\item{size}{The number of draws}

\item{mu}{The mean proportion.}

\item{rho}{The overdispersion parameter.}

\item{return_log}{A logical. Should we return the log density
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
double version of betabinomial density.
}
\author{
David Gerard
}
