% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deps_dell}
\alias{deps_dell}
\title{Derivative of exp(ell) / (1 + exp(ell))}
\usage{
deps_dell(ell)
}
\arguments{
\item{ell}{A double.}
}
\description{
exp(ell) / (1 + exp(ell)) is the sequencing error rate and ell is a parameterization to
that is unconstrained.
}
\author{
David Gerard
}
