% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{df_deps}
\alias{df_deps}
\title{Returns derivative of p(1 - eps) + (1 - p) * eps}
\usage{
df_deps(eps, p)
}
\arguments{
\item{eps}{The sequencing error rate.}

\item{p}{The proportion of genomes whose allele is A.}
}
\description{
Returns derivative of p(1 - eps) + (1 - p) * eps
}
\author{
David Gerard
}
