% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dh_dtau}
\alias{dh_dtau}
\title{Derivative of h(tau) = (1 - tau) / tau.}
\usage{
dh_dtau(tau)
}
\arguments{
\item{tau}{A double. The overdispersion parameter}
}
\description{
Just returns -1 / tau^2
}
\author{
David Gerard
}
