% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dhyper_cpp}
\alias{dhyper_cpp}
\title{Vectorized version of \code{\link[stats]{dhyper}} for C++ implementation.}
\usage{
dhyper_cpp(x, m, n, k)
}
\arguments{
\item{x}{Vector of quantiles representing the number of white balls drawn without replacement from an urn which contains both black and white balls.}

\item{m}{The number of white balls in the urn.}

\item{n}{The number of black balls in the urn.}

\item{k}{The number of balls drawn from the urn.}
}
\description{
Vectorized version of \code{\link[stats]{dhyper}} for C++ implementation.
}
\author{
David Gerard
}
