% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{dupdog}
\alias{dupdog}
\title{Returns log-likelihood of \code{updog} object.}
\usage{
dupdog(obj)
}
\arguments{
\item{obj}{An object of class \code{updog} returned from \code{\link{updog}}.}
}
\description{
Returns log-likelihood of \code{updog} object.
}
\author{
David Gerard
}
