% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dxi_dd}
\alias{dxi_dd}
\title{Derivative w.r.t. d of xi(d, f) = f / (d * (1 - f) + f)}
\usage{
dxi_dd(d, f)
}
\arguments{
\item{d}{The bias term. Must be greater than 0.}

\item{f}{The sequencing-error adjusted probability of the refernence allele.}
}
\description{
Derivative w.r.t. d of xi(d, f) = f / (d * (1 - f) + f)
}
\author{
David Gerard
}
