% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_seq.R
\name{est_seq_post}
\alias{est_seq_post}
\title{Use posterior probabilities to update the estimated sequencing error.}
\usage{
est_seq_post(opostprob, ocounts, osize)
}
\arguments{
\item{opostprob}{The matrix output from up_max_bb}

\item{ocounts}{The offspring counts of the reference allele.}

\item{osize}{The offspring total reads.}
}
\value{
The updated estimate of the sequencing error rate.
}
\description{
Use posterior probabilities to update the estimated sequencing error.
}
\author{
David Gerard
}

