% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{get_cov_mle}
\alias{get_cov_mle}
\title{Calculate the negative inverse hessian of the parameters from
\code{\link{obj_offspring_reparam}}.}
\usage{
get_cov_mle(ocounts, osize, ploidy, prob_geno, bias_val, seq_error, od_param,
  out_prop, p1counts = NULL, p1size = NULL, p1geno = NULL,
  p2counts = NULL, p2size = NULL, p2geno = NULL, bias_val_mean = 0,
  bias_val_sd = 1, seq_error_mean = -4.7, seq_error_sd = 1,
  model = c("f1", "s1", "hw", "uniform"))
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{prob_geno}{The allele frequencies of the genotypes. See \code{\link{get_prob_geno}}.}

\item{bias_val}{The bias parameter. A value of 1 means there is no bias
toward one allele or the other. A value less than one indicates a bias
toward the reference allele. A value greater than one indicates a bias
toward the non-reference allele.}

\item{seq_error}{The sequencing error rate.}

\item{od_param}{The overdispersion parameter in the beta-binomial model
for the OK counts. When this is zero, this resorts to the binomial
model for counts.}

\item{out_prop}{The proportion of points that are outliers. Defaults
(quite arbitrarily) to \code{0.01}.}

\item{p1counts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith
sample of parent 1.  If \code{NULL} then the prior
probabilities on parent 1's genotype will default to uniform.}

\item{p1size}{A vector of positive integers. The ith element is the
total number of reads in the ith sample of parent 1.  If
\code{NULL} then the prior probabilities on parent 1's genotype
will default to uniform.}

\item{p1geno}{The initial value of the first parental genotype.}

\item{p2counts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith
sample of parent 2.  If \code{NULL} then the prior
probabilities on parent 2's genotype will default to uniform.}

\item{p2size}{A vector of positive integers. The ith element is the
total number of reads in the ith sample of parent 2.  If
\code{NULL} then the prior probabilities on parent 2's genotype
will default to uniform.}

\item{p2geno}{The initial value of the second parental genotype.}

\item{bias_val_mean}{The prior mean on the log of \code{bias_val} (corresponding to \code{parvec[1]}).
Set \code{bias_val_sd = Inf} to have no penalty on the bias parameter.}

\item{bias_val_sd}{The prior standard deviation on the log of \code{bias_val}
(corresponding to \code{parvec[1]}).
Set \code{bias_val_sd = Inf} to have no penalty on the bias parameter.}

\item{seq_error_mean}{The mean of the logit-normal prior on the sequencing error rate, which corresponds
to \code{parvec[2]}. Set \code{seq_error_sd = Inf} to have no penalty on the sequencing error rate.
The default is -4.7, which roughly corresponds to a mean sequencing error value of 0.009.}

\item{seq_error_sd}{The standard deviation of the logit-normal prior on the sequencing error rate, which
corresponds to \code{parvec[2]}. The default is 1, which at three standard deviations is about a sequencing
error rate of 0.15.}

\item{model}{The model for the genotype distribution. Do we assume an
F1 population (\code{"f1"}), an S1 population (\code{"s1"}), Hardy-Weinberg equilibrium (\code{"hw"}),
or a uniform distribution (\code{"uniform"}).}
}
\description{
This is the covariance matrix if evaluated at the MLE's of these parameters.
}
\author{
David Gerard
}
