% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_out_prop}
\alias{get_out_prop}
\title{E step in EM algorithm for a single individual.}
\usage{
get_out_prop(ocounts, osize, ploidy, prob_geno, d, eps, tau, out_prop, out_mean,
  out_disp)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{prob_geno}{The allele frequencies of the genotypes. See \code{\link{get_prob_geno}}.}

\item{d}{The sequencing bias parameter.}

\item{eps}{The sequencing error rate}

\item{tau}{The overdisperion parameter.}

\item{out_prop}{The proportion of points that are outliers. Defaults
(quite arbitrarily) to \code{0.01}.}

\item{out_mean}{The mean of beta-binomial for the outlier distribution.
Defaults to \code{0.5}.}

\item{out_disp}{The overdispersion parameter of the outlier distribution.
Defaults to \code{1/3}, which corresponds to a uniform distribution
for the underlying beta when \code{out_mean = 0.5}.}
}
\description{
E step in EM algorithm for a single individual.
}
\author{
David Gerard
}
