% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_parent_outprop}
\alias{get_parent_outprop}
\title{E-step for the parents.}
\usage{
get_parent_outprop(pcounts, psize, ploidy, pgeno, d, eps, tau, out_prop,
  out_mean, out_disp)
}
\arguments{
\item{pcounts}{The number of counts of the reference allele we observe in
a parent.}

\item{psize}{The total number of reads we observe in a parent.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{pgeno}{The genotype of the parent.}

\item{d}{Same as \code{bias_val} in \code{\link{obj_parent}}.}

\item{eps}{Same as \code{seq_error} in \code{\link{obj_parent}}.}

\item{tau}{Same as \code{od_param} in \code{\link{obj_parent}}.}

\item{out_prop}{The proportion of points that are outliers. Defaults
(quite arbitrarily) to \code{0.01}.}

\item{out_mean}{The mean of beta-binomial for the outlier distribution.
Defaults to \code{0.5}.}

\item{out_disp}{The overdispersion parameter of the outlier distribution.
Defaults to \code{1/3}, which corresponds to a uniform distribution
for the underlying beta when \code{out_mean = 0.5}.}
}
\description{
E-step for the parents.
}
\author{
David Gerard
}
