% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_prob_geno}
\alias{get_prob_geno}
\title{Rcpp function to get genotype probabilities assuming either
F1 population given parental genotypes, the allele frequencies assuming
Hardy-Weinberg equilibrium, or just a uniform distribution.}
\usage{
get_prob_geno(ploidy, model, p1geno, p2geno, allele_freq)
}
\arguments{
\item{ploidy}{The ploidy of the species.}

\item{model}{Do we assume the genotypes are distributed from an F1 population (\code{"f1"}),
according to Hardy-Weinberg (\code{hw}), or uniformly (\code{"uniform"})?}

\item{p1geno}{The first parental genotype if \code{model = "f1"}.}

\item{p2geno}{The second parental genotype if \code{model = "f1"}.}

\item{allele_freq}{The allele-frequency if \code{model = "hw"}.}
}
\description{
Rcpp function to get genotype probabilities assuming either
F1 population given parental genotypes, the allele frequencies assuming
Hardy-Weinberg equilibrium, or just a uniform distribution.
}
\author{
David Gerard
}
