% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pvec}
\alias{get_pvec}
\title{Gets all possible binomial probabilities for a given ploidy, bias term, and sequencing
error rate.}
\usage{
get_pvec(ploidy, bias_val, seq_error)
}
\arguments{
\item{ploidy}{The ploidy of the species.}

\item{bias_val}{The bias parameter. Should be greater than 0. A value of 1 means no bias.}

\item{seq_error}{The sequencing error rate.}
}
\description{
Gets all possible binomial probabilities for a given ploidy, bias term, and sequencing
error rate.
}
\author{
David Gerard
}
