% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_q_array_cpp}
\alias{get_q_array_cpp}
\title{Rcpp implementation of \code{\link{get_q_array}}.}
\usage{
get_q_array_cpp(ploidy)
}
\arguments{
\item{ploidy}{A positive integer. The ploidy of the species.}
}
\description{
This function will return the segregation proabilities
}
\seealso{
\code{\link{get_q_array}}.
}
\author{
David Gerard
}
