% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog.R
\name{get_transition_mat}
\alias{get_transition_mat}
\title{Get the transition probabilities from going from one ploidy to
another assuming a constant error rate.}
\usage{
get_transition_mat(ploidy, eps)
}
\arguments{
\item{ploidy}{The ploidy of the species.}

\item{eps}{The probability of one allele switching.}
}
\description{
Get the transition probabilities from going from one ploidy to
another assuming a constant error rate.
}
\author{
David Gerard
}

