% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_bb.R
\name{good_bb_obj}
\alias{good_bb_obj}
\title{Objective function for non-outlier model}
\usage{
good_bb_obj(rho, ocounts, osize, theta, avec, pk)
}
\arguments{
\item{rho}{The overdispersion parameter.}

\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{theta}{A vector of proportions. The posterior probabilities of each point being ok.}

\item{avec}{The current slice of \code{q_array}.}

\item{pk}{The parameterspace.}
}
\description{
Objective function for non-outlier model
}
\author{
David Gerard
}

