% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_offspring}
\alias{grad_offspring}
\title{Gradient of \code{\link{obj_offspring_reparam}}.}
\usage{
grad_offspring(ocounts, osize, ploidy, prob_geno, s, ell, r)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{prob_geno}{The allele frequencies of the genotypes. See \code{\link{get_prob_geno}}.}

\item{s}{We have \code{s = exp(d)}, where \code{exp(d)} is the bias term.}

\item{ell}{The logit of the sequencing error rate}

\item{r}{We have \code{r = log((1 - tau) / tau)}}
}
\description{
Gradient of \code{\link{obj_offspring_reparam}}.
}
\author{
David Gerard
}
