% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_offspring_mat}
\alias{grad_offspring_mat}
\title{Gradient of \code{\link{obj_offspring}} for each individual.}
\usage{
grad_offspring_mat(ocounts, osize, ploidy, p1geno, p2geno, s, ell, r)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{p1geno}{The first parental genotype. The number of copies of the
reference allele the first parent has.}

\item{p2geno}{The second parental genotype. The number of copies of the
reference allele the second parent has.}

\item{s}{We have \code{s = exp(d)}, where \code{exp(d)} is the bias term.}

\item{ell}{The logit of the sequencing error rate}

\item{r}{We have \code{r = log((1 - tau) / tau)}}
}
\description{
Gradient of \code{\link{obj_offspring}} for each individual.
}
\author{
David Gerard
}
