% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_offspring_weights_original}
\alias{grad_offspring_weights_original}
\title{Gradient of \code{\link{obj_offspring_weights}} using original parameterization}
\usage{
grad_offspring_weights_original(ocounts, osize, weight_vec, ploidy, prob_geno,
  d, eps, tau)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{weight_vec}{A vector of weights between 0 and 1 (do not need to add up to 1).}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{prob_geno}{The allele frequencies of the genotypes. See \code{\link{get_prob_geno}}.}

\item{d}{The sequencing bias parameter.}

\item{eps}{The sequencing error rate}

\item{tau}{The overdisperion parameter.}
}
\value{
A NumericVector of length three with the partial derivatives of
    \code{d}, \code{eps}, and \code{tau}, in that order.
}
\description{
Gradient of \code{\link{obj_offspring_weights}} using original parameterization
}
