% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_parent}
\alias{grad_parent}
\title{Gradient of \code{\link{obj_parent}} (when \code{outlier = FALSE}) with respect to
\code{bias_val}, \code{seq_error}, and \code{od_param}. Basically, this just calculates
derivatives of the log beta-binomial density.}
\usage{
grad_parent(pcounts, psize, ploidy, pgeno, bias_val = 1, seq_error = 0,
  od_param = 0, weight = 1)
}
\arguments{
\item{pcounts}{The number of counts of the reference allele we observe in
a parent.}

\item{psize}{The total number of reads we observe in a parent.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{pgeno}{The genotype of the parent.}

\item{bias_val}{The bias parameter. A value of 1 means there is no bias
toward one allele or the other. A value less than one indicates a bias
toward the reference allele. A value greater than one indicates a bias
toward the non-reference allele.}

\item{seq_error}{The sequencing error rate.}

\item{od_param}{The overdispersion parameter in the beta-binomial model
for the OK counts. When this is zero, this resorts to the binomial
model for counts.}

\item{weight}{A double that multiplies to the log-likelihood (objective).
Should only be used when \code{outlier = FALSE} during the M
step of the EM algorithm.}
}
\value{
A \code{NumericVector} of length three that contains the partial derivatives of
    \code{bias_val}, \code{seq_error}, and \code{od_param}, in that order.
}
\description{
Gradient of \code{\link{obj_parent}} (when \code{outlier = FALSE}) with respect to
\code{bias_val}, \code{seq_error}, and \code{od_param}. Basically, this just calculates
derivatives of the log beta-binomial density.
}
\author{
David Gerard
}
