% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_parent_reparam}
\alias{grad_parent_reparam}
\title{The gradient of \code{\link{obj_parent_reparam}} with respect to \code{s}, \code{ell},
and \code{r}, which recall are reparameterizations of the bias, seqeuncing error, and
overdispersion, respectively.}
\usage{
grad_parent_reparam(pcounts, psize, ploidy, pgeno, s, ell, r, weight = 1)
}
\arguments{
\item{pcounts}{The number of counts of the reference allele we observe in
a parent.}

\item{psize}{The total number of reads we observe in a parent.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{pgeno}{The genotype of the parent.}

\item{s}{Same as \code{log(bias_val)} in \code{\link{obj_offspring}}.}

\item{ell}{We have \code{seq_error = expit(ell)} from \code{\link{obj_offspring}}.}

\item{r}{Same as \code{log((1.0 - od_param) / od_param)} from \code{\link{obj_offspring}}.}

\item{weight}{A double that multiplies to the log-likelihood (objective).
Should only be used when \code{outlier = FALSE} during the M
step of the EM algorithm.}
}
\value{
A \code{NumericVector} of length three with the partial derivatives of \code{s},
    \code{ell}, and \code{r} in that order.
}
\description{
The gradient of \code{\link{obj_parent_reparam}} with respect to \code{s}, \code{ell},
and \code{r}, which recall are reparameterizations of the bias, seqeuncing error, and
overdispersion, respectively.
}
\author{
David Gerard
}
