% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.updog}
\alias{is.updog}
\title{Tests if its argument is an updog.}
\usage{
is.updog(x)
}
\arguments{
\item{x}{Anything.}
}
\value{
A logical. \code{TRUE} if \code{x} is an updog, and \code{FALSE} otherwise.
}
\description{
Tests if its argument is an updog.
}
\author{
David Gerard
}

