% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logsumexp}
\alias{logsumexp}
\title{Log-sum-exponential trick that I use all the time.}
\usage{
logsumexp(xx)
}
\arguments{
\item{xx}{A matrix whose rows are to be log-sum-exponentiated.}
}
\description{
Log-sum-exponential trick that I use all the time.
}
\author{
David Gerard
}
