% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_offspring}
\alias{obj_offspring}
\title{Objective function for the offspring.}
\usage{
obj_offspring(ocounts, osize, ploidy, prob_geno, bias_val = 1,
  seq_error = 0, od_param = 0, outlier = FALSE, out_prop = 0.01,
  out_mean = 0.5, out_disp = 1/3)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{prob_geno}{The allele frequencies of the genotypes. See \code{\link{get_prob_geno}}.}

\item{bias_val}{The bias parameter. A value of 1 means there is no bias
toward one allele or the other. A value less than one indicates a bias
toward the reference allele. A value greater than one indicates a bias
toward the non-reference allele.}

\item{seq_error}{The sequencing error rate.}

\item{od_param}{The overdispersion parameter in the beta-binomial model
for the OK counts. When this is zero, this resorts to the binomial
model for counts.}

\item{outlier}{A logical. Should we include an outlier model (\code{TRUE})
or not (\code{FALSE})? Defaults to \code{FALSE}.}

\item{out_prop}{The proportion of points that are outliers. Defaults
(quite arbitrarily) to \code{0.01}.}

\item{out_mean}{The mean of beta-binomial for the outlier distribution.
Defaults to \code{0.5}.}

\item{out_disp}{The overdispersion parameter of the outlier distribution.
Defaults to \code{1/3}, which corresponds to a uniform distribution
for the underlying beta when \code{out_mean = 0.5}.}
}
\description{
Objective function for the offspring.
}
\author{
David Gerard
}
