% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_offspring_reparam}
\alias{obj_offspring_reparam}
\title{Just a reparameterization of \code{\link{obj_offspring}}.}
\usage{
obj_offspring_reparam(ocounts, osize, ploidy, p1geno, p2geno, d, ell, h)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{p1geno}{The first parental genotype. The number of copies of the
reference allele the first parent has.}

\item{p2geno}{The second parental genotype. The number of copies of the
reference allele the second parent has.}

\item{d}{Same as \code{bias_val} in \code{\link{obj_offspring}}.}

\item{ell}{We have \code{seq_error = expit(ell)} from \code{\link{obj_offspring}}.}

\item{h}{Same as \code{(1.0 - od_param) / od_param} from \code{\link{obj_offspring}}.}
}
\description{
Just a reparameterization of \code{\link{obj_offspring}}.
}
\author{
David Gerard
}
