% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_parent_reparam}
\alias{obj_parent_reparam}
\title{A reparameterization of \code{\link{obj_parent}}}
\usage{
obj_parent_reparam(pcounts, psize, ploidy, pgeno, s, ell, r, weight = 1)
}
\arguments{
\item{pcounts}{The number of counts of the reference allele we observe in
a parent.}

\item{psize}{The total number of reads we observe in a parent.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{pgeno}{The genotype of the parent.}

\item{s}{Same as \code{exp(bias_val)} in \code{\link{obj_offspring}}.}

\item{ell}{We have \code{seq_error = expit(ell)} from \code{\link{obj_offspring}}.}

\item{r}{Same as \code{log((1.0 - od_param) / od_param)} from \code{\link{obj_offspring}}.}

\item{weight}{A double between 0 and 1. This should only not be 1 when
doing the em algorithm and the weight is the probabiility that
a point being ok.}
}
\description{
A reparameterization of \code{\link{obj_parent}}
}
\author{
David Gerard
}
