% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{obj_wrapp_all}
\alias{obj_wrapp_all}
\title{wrapper for \code{\link{obj_offspring_weights_reparam}}}
\usage{
obj_wrapp_all(parvec, ocounts, osize, weight_vec, ploidy, p1geno, p2geno,
  bound_bias = TRUE, bound_od = TRUE, update_bias_val = TRUE,
  update_seq_error = TRUE, update_od_param = TRUE)
}
\arguments{
\item{parvec}{A vector of three elements, s, ell, and r. We have s = log(bias_val) = log(d),
ell = logit(seq_error) = logit(eps), and r = 1 / logit(od_param) = 1 / logit(tau).}

\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{weight_vec}{A vector of weights obtained via the E-step.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{p1geno}{The first parental genotype. The number of copies of the
reference allele the first parent has.}

\item{p2geno}{The second parental genotype. The number of copies of the
reference allele the second parent has.}

\item{bound_bias}{A logical. Should we bound the bias parameter \code{parvec[1]} by a somewhat arbitrary value
(\code{TRUE}) or not (\code{FALSE})?}

\item{bound_od}{A logical. Should we bound the overdisperion parameter
\code{parvec[3]} by a very small number (\code{TRUE}) or not
(\code{FALSE}). This is mostly because my code for super super small
values of \code{parvec[3]} can be unstable.}

\item{update_bias_val}{A logical. Not used. Here for compatability
with \code{\link{grad_wrapp_all}}.}

\item{update_seq_error}{A logical. Not used. Here for compatability
with \code{\link{grad_wrapp_all}}.}

\item{update_od_param}{A logical. Not used. Here for compatability
with \code{\link{grad_wrapp_all}}.}
}
\description{
wrapper for \code{\link{obj_offspring_weights_reparam}}
}
\author{
David Gerard
}
