% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{out_grad_parent}
\alias{out_grad_parent}
\title{Gradient function for \code{\link{dbetabinom_mu_rho_cpp_double}}.}
\usage{
out_grad_parent(pcounts, psize, out_mean, out_disp, weight)
}
\arguments{
\item{pcounts}{The number of counts of the reference allele we observe in
a parent.}

\item{psize}{The total number of reads we observe in a parent.}

\item{out_mean}{The mean of the BB.}

\item{out_disp}{The overdispersion parameter of the BB.}

\item{weight}{The probability a point is an outlier.}
}
\value{
A vector of length two. The first element of which is the partial derivative
    of the outlier mean. The second element of which is the partial
    derivative of the outlier overdispersion parameter.
}
\description{
Gradient function for \code{\link{dbetabinom_mu_rho_cpp_double}}.
}
\author{
David Gerard
}
