% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_bb.R
\name{outlier_bb_obj}
\alias{outlier_bb_obj}
\title{Objective function for outlier model.}
\usage{
outlier_bb_obj(par, ocounts, osize, theta)
}
\arguments{
\item{par}{A 2-vector. The first element is the mean of the
beta-binomial. The second is the overdispersion parameter.}

\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{theta}{A vector of proportions. The posterior probabilities of each point being ok.}
}
\description{
Objective function for outlier model.
}
\author{
David Gerard
}

