% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outlier_obj}
\alias{outlier_obj}
\title{Objective function of outlier part in EM step}
\usage{
outlier_obj(ocounts, osize, weight_vec, out_mean, out_disp)
}
\arguments{
\item{ocounts}{A vector of reads counts of reference allele.}

\item{osize}{A vector of total reads counts.}

\item{weight_vec}{The current probability of a point being an outlier.}

\item{out_mean}{The current mean of the beta.}

\item{out_disp}{The current overdispersion parameter of the beta.}
}
\description{
Objective function of outlier part in EM step
}
\author{
David Gerard
}
