% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbias}
\alias{pbias}
\title{Returns the probability of seeing the reference allele after including
the mapping-bias and the sequencing-error.}
\usage{
pbias(prob, bias, seq_error)
}
\arguments{
\item{prob}{A numeric vector. Each element is the proportion of genomes that contain
the reference allele. This should take on
one of the values 0/K, 1/K, ... , K/K, where K is the ploidy of the individual.}

\item{bias}{The bias parameter. Should be greater than or equal to zero,
though is typically less than 1.}

\item{seq_error}{The sequencing error rate. This should be between 0 and 1.}
}
\description{
Returns the probability of seeing the reference allele after including
the mapping-bias and the sequencing-error.
}
\author{
David Gerard
}
