% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbias_double}
\alias{pbias_double}
\title{A double version of \code{\link{pbias}}.}
\usage{
pbias_double(prob, bias, seq_error)
}
\arguments{
\item{prob}{A numeric vector. Each element is the proportion of genomes that contain
the reference allele. This should take on
one of the values 0/K, 1/K, ... , K/K, where K is the ploidy of the individual.}

\item{bias}{The bias parameter. Should be greater than or equal to zero,
though is typically less than 1. A 1 indicates no bias. A value less than one
indicates a bias towards the reference allele. A value greater than 1 indicates
a bias towards the non-reference allele.}

\item{seq_error}{The sequencing error rate. This should be between 0 and 1.}
}
\description{
A double version of \code{\link{pbias}}.
}
\seealso{
\code{\link{pbias}}.
}
\author{
David Gerard
}
