% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.updog}
\alias{plot.updog}
\title{Draw a genotype plot from the output of \code{\link{updog}}.}
\usage{
\method{plot}{updog}(x, gg = requireNamespace("ggplot2", quietly = TRUE),
  plot_beta = TRUE, ...)
}
\arguments{
\item{x}{The output from \code{\link{updog}}.}

\item{gg}{Should we use ggplot2 to plot the genotypes
(\code{TRUE}), or not (\code{FALSE}).  If ggplot2 is present, then
this defaults to \code{TRUE}. If it is not present, then it
defaults to \code{FALSE}.}

\item{plot_beta}{A logical. If true, then we'll also plot the
estimated beta density of the outlier model, followed by the
estimated beta distributions of the overdispersion models.}

\item{...}{Not used.}
}
\value{
A plot object.
}
\description{
Iff ggplot2 is installed, then this function will use it to make
the genotype plot.  Otherwise, "classic" R base graphics will be
used. The ggplot2 version is made using the function
\code{\link{plot_geno}}.
}
\author{
David Gerard
}

