% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_beta_dist}
\alias{plot_beta_dist}
\title{Plot the beta distribution.}
\usage{
plot_beta_dist(alpha = NULL, beta = NULL, mu = NULL, rho = NULL, ...)
}
\arguments{
\item{alpha}{The shape1 parameter.}

\item{beta}{The shape2 parameter.}

\item{mu}{The mean parameter.}

\item{rho}{The overdispersion parameter.}

\item{...}{Anything else you want to pass to \code{\link[graphics]{plot.default}}.}
}
\description{
This function will plot the beta distribution under two different
parameterizations. The first one, with \code{alpha} and \code{beta},
is the usual parameterization of the beta. The second one parameterizes
the beta in terms of its mean and overdispersion. Thus,
either both \code{alpha} and \code{beta} must be specified,
or both \code{mu} and \code{rho} must be specified.
}
\details{
We have the relationships
\deqn{\mu = \alpha / (\alpha + \beta)}
and
\deqn{\rho = 1 / (1 + \alpha + \beta).}
Thus, the inverse relationships are
\deqn{\alpha = \mu(1 -\rho) / \rho}
and
\deqn{\beta = (1 - \mu)(1 - \rho) / \rho.}
}
\author{
David Gerard
}
