% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_geno_base}
\alias{plot_geno_base}
\title{The base R graphics version of \code{\link{plot_geno}}.}
\usage{
plot_geno_base(ocounts, osize, ploidy, p1counts = NULL, p1size = NULL,
  p2counts = NULL, p2size = NULL, ogeno = NULL, seq_error = 0,
  bias_val = 1, prob_ok = NULL, maxpostprob = NULL, p1geno = NULL,
  p2geno = NULL)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The number of
reference alleles observed in the offspring.}

\item{osize}{A vector of positive integers. The total number of
reads in the offspring.}

\item{ploidy}{A non-negative integer. The ploidy of the species.}

\item{p1counts}{A vector of non-negative integers. The number of
reference alleles observed in parent 1.}

\item{p1size}{A vector of positive integers. The total number of
reads in parent 1.}

\item{p2counts}{A vector of non-negative integers. The number of
reference alleles observed in parent 2.}

\item{p2size}{A vector of positive integers. The total number of
reads in parent 2.}

\item{ogeno}{The child genotypes}

\item{seq_error}{The average sequencing error rate.}

\item{bias_val}{The bias parameter.}

\item{prob_ok}{A vector of posterior probabilities of not being a mistake.}

\item{maxpostprob}{A vector of the posterior probabilities of begin at the modal probability.}

\item{p1geno}{Parent 1's genotype.}

\item{p2geno}{Parent 2's genotype.}
}
\description{
The base R graphics version of \code{\link{plot_geno}}.
}
\seealso{
\code{\link{plot_geno}}
}
\author{
David Gerard
}
