% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plot.R
\name{plot_problines}
\alias{plot_problines}
\title{Plots the genotype lines for a given sequencing error rate, bias parameter, and ploidy.}
\usage{
plot_problines(seq_error, bias_val, ploidy)
}
\arguments{
\item{seq_error}{The sequencing error rate.}

\item{bias_val}{The bias parameter.}

\item{ploidy}{The ploidy.}
}
\description{
Plots the genotype lines for a given sequencing error rate, bias parameter, and ploidy.
}
\author{
David Gerard
}
