% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{rbetabinom_mu_rho}
\alias{rbetabinom_mu_rho}
\title{Draw from beta-binomial distribution parameterized
by mean and overdispersion parameter.}
\usage{
rbetabinom_mu_rho(mu, rho, size)
}
\arguments{
\item{mu}{The mean of the underlying beta distribution. This is
equal to \eqn{a / (a + b)} where \eqn{a} and \eqn{b} are the two
shape parameters of the usual parameterization of the beta. This can
be a vector of values.}

\item{rho}{The over-dispersion parameter of the underlying beta
distribution. This is equal to \eqn{1 / (a + b + 1)} where \eqn{a}
and \eqn{b} are the two shape parameters of the usual parameterization
of the beta. This can be a vector of values.}

\item{size}{The sample size. This can equal a vector of observations.}
}
\description{
Draw from beta-binomial distribution parameterized
by mean and overdispersion parameter.
}
\author{
David Gerard
}
