% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{rupdog}
\alias{rupdog}
\title{Sample counts using the parameter values in an \code{updog} object.}
\usage{
rupdog(obj)
}
\arguments{
\item{obj}{An object of class \code{updog} returned from \code{\link{updog}}.}
}
\description{
Sample counts using the parameter values in an \code{updog} object.
}
\author{
David Gerard
}
