% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{screen_oracle}
\alias{screen_oracle}
\title{Returns the distribution of proportion of individuals genotyped correctly using an the oracle estimator
when the parameter values equal the estimated values.}
\usage{
screen_oracle(obj, itermax = 1000)
}
\arguments{
\item{obj}{An object of class \code{updog}.}

\item{itermax}{The number of repetitions to draw.}
}
\value{
A vector of length \code{itermax} which contains samples from the distribution of proportion of
    individuals genotyped correctly by an oracle estimator.
}
\description{
Returns the distribution of proportion of individuals genotyped correctly using an the oracle estimator
when the parameter values equal the estimated values.
}
\author{
David Gerard
}
