% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_bb.R
\name{up_bb_fix}
\alias{up_bb_fix}
\title{EM iteration for overdispersed binomial mixture model.}
\usage{
up_bb_fix(pival, p1geno, p2geno, rho, out_mu, out_rho, ocounts, osize, qarray,
  r1vec, r2vec, pk, update_pival = TRUE, update_rho = TRUE,
  update_geno = FALSE, update_outlier = TRUE, upper_rho = NULL)
}
\arguments{
\item{pival}{A fudge factor. The probability of not begin a
mistake. Estimated if \code{est_fudge = TRUE}. Set to 0.99. Only used if \code{est_fudge = TRUE}.}

\item{p1geno}{The current genotype of parent 1. Must be in \code{0:ploidy}.}

\item{p2geno}{The current genotype of parent 1. Must be in \code{0:ploidy}.}

\item{rho}{The overdispersion parameter.}

\item{out_mu}{The outlier mean.}

\item{out_rho}{The outlier overdispersion parameter.}

\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{qarray}{An three-way array of proportions. The (i, j, k)th
element is the probability of an offspring having k - 1
reference alleles given that parent 1 has i - 1 refrerence
alleles and parent 2 has j - 1 reference alleles. Each
dimension of the array is \code{ploidy + 1}.}

\item{r1vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 1 has i -1 copies of allele
A.}

\item{r2vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 2 has i -1 copies of allele
A.}

\item{pk}{The parameter space. Should all be between 0 and 1.}

\item{update_pival}{A logical. Should we update the mixture proportion
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_rho}{A logical. Should we update the overdispersion
parameters (\code{TRUE}) or not (\code{FALSE})?}

\item{update_geno}{A logical. Should we update the parental
genotypes (\code{TRUE}) or not (\code{FALSE})?}

\item{update_outlier}{A logical. Should we update the outlier
distribution (\code{TRUE}) or not (\code{FALSE})?}

\item{upper_rho}{The upperbound on the overdispersion parameter. Defaults to 1 / (ploidy + 1) which corrsponds
to bounding smallest probabilities to have finite density at 0 and 1.}
}
\description{
EM iteration for overdispersed binomial mixture model.
}
\author{
David Gerard
}
