% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_bb.R
\name{up_bb_obj}
\alias{up_bb_obj}
\title{Objective function for overdispersed binomial mixture model}
\usage{
up_bb_obj(pival, p1geno, p2geno, rho, out_mu, out_rho, ocounts, osize, qarray,
  r1vec, r2vec, pk)
}
\arguments{
\item{pival}{A fudge factor. The probability of not begin a
mistake. Estimated if \code{est_fudge = TRUE}. Set to 0.99. Only used if \code{est_fudge = TRUE}.}

\item{p1geno}{The genotype of parent 1. From 0 to ploidy.}

\item{p2geno}{The genotype of parent 2. From 0 to ploidy.}

\item{rho}{The overdispersion parameter.}

\item{out_mu}{The outlier mean.}

\item{out_rho}{The outlier overdispersion parameter.}

\item{ocounts}{A vector of non-negative integers. The observed number of A's.}

\item{osize}{A vector of positive integers. The observed total counts.}

\item{qarray}{An three-way array of proportions. The (i, j, k)th
element is the probability of an offspring having k - 1
reference alleles given that parent 1 has i - 1 refrerence
alleles and parent 2 has j - 1 reference alleles. Each
dimension of the array is \code{ploidy + 1}.}

\item{r1vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 1 has i -1 copies of allele
A.}

\item{r2vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 2 has i -1 copies of allele
A.}

\item{pk}{The parameter space. Should all be between 0 and 1.}
}
\description{
Objective function for overdispersed binomial mixture model
}
\author{
David Gerard
}

