% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog.R
\name{up_fix}
\alias{up_fix}
\title{Fixed-point iteration for updog EM.}
\usage{
up_fix(pival, p1geno, p2geno, alpha, beta, ocounts, osize, ldbinommat, qarray,
  r1vec, r2vec, update_pi = TRUE, update_geno = TRUE, update_beta = TRUE)
}
\arguments{
\item{pival}{A fudge factor. The probability of not begin a
mistake. Estimated if \code{est_fudge = TRUE}. Set to 0.99. Only used if \code{est_fudge = TRUE}.}

\item{p1geno}{The genotype of parent 1}

\item{p2geno}{The genotype of parent 2}

\item{alpha}{The alpha parameter in the beta distribution of the outliers. Only used if \code{est_fudge = TRUE}.}

\item{beta}{The beta parameter in the beta distribution of the outliers. Only used if \code{est_fudge = TRUE}.}

\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{ldbinommat}{The log binomial densities of the counts. The rows index the
parental genotypes and the columns index the individuals.}

\item{qarray}{An three-way array of proportions. The (i, j, k)th
element is the probability of an offspring having k - 1
reference alleles given that parent 1 has i - 1 refrerence
alleles and parent 2 has j - 1 reference alleles. Each
dimension of the array is \code{ploidy + 1}.}

\item{r1vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 1 has i -1 copies of allele
A.}

\item{r2vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 2 has i -1 copies of allele
A.}

\item{update_pi}{A logical. Should we update the fudge factor
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_geno}{A logical. Should we update the parental
genotypes (\code{TRUE}) or not (\code{FALSE})?}

\item{update_beta}{A logical. Should we update the outlier model
\code{TRUE} or not (\code{FALSE})?}
}
\description{
Fixed-point iteration for updog EM.
}
\author{
Davod Gerard
}
