% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_bb.R
\name{up_max_bb}
\alias{up_max_bb}
\title{EM algorithm for over-dispersed mixture model.}
\usage{
up_max_bb(ocounts, osize, qarray, r1vec, r2vec, pk, pival = 0.99,
  out_mu = 0.5, out_rho = 0.8, rho = 0.001, est_fudge = TRUE,
  tol = 10^-4, itermax = 1000, update_geno = FALSE,
  update_pival = FALSE, update_outlier = FALSE, update_rho = FALSE)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{qarray}{An three-way array of proportions. The (i, j, k)th
element is the probability of an offspring having k - 1
reference alleles given that parent 1 has i - 1 refrerence
alleles and parent 2 has j - 1 reference alleles. Each
dimension of the array is \code{ploidy + 1}.}

\item{r1vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 1 has i -1 copies of allele
A.}

\item{r2vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 2 has i -1 copies of allele
A.}

\item{pk}{The parameter space. Should all be between 0 and 1.}

\item{pival}{A fudge factor. The probability of not begin a
mistake. Estimated if \code{est_fudge = TRUE}. Set to 0.99. Only used if \code{est_fudge = TRUE}.}

\item{out_mu}{The outlier mean.}

\item{out_rho}{The outlier overdispersion parameter.}

\item{rho}{The overdispersion parameter. Needs to be between 0 and 1.
If it is exactly 0, you should be using \code{\link{updog_maximize}}.}

\item{est_fudge}{A logical. Should we estimate the fudge factor
\code{TRUE} or not \code{FALSE}?}

\item{tol}{A positive numeric. The tolerance for the stopping criterion.}

\item{itermax}{A positive integer. The maximum number of iterations to run in the EM.}

\item{update_geno}{A logical. Update the parental genotypes?}

\item{update_pival}{A logical. Should we update the mixing proportion (\code{TRUE}) or not (\code{FALSE})?}

\item{update_outlier}{A logical. Should we update the outlier distribution (\code{TRUE}) or not (\code{FALSE})?}

\item{update_rho}{A logical. Should we we update the overdispersion parameter (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
EM algorithm for over-dispersed mixture model.
}
\author{
David Gerard
}
