% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog.R
\name{updog}
\alias{updog}
\title{Using Parental Data for Offspring Genotyping.}
\usage{
updog(ocounts, osize, ploidy, model = c("f1", "s1", "hw", "uniform"), ...)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{ploidy}{A positive integer. The number of copies of the
genome in the species.}

\item{model}{The model for the genotype distribution. Do we assume an
F1 population (\code{"f1"}), an S1 population (\code{"s1"}), Hardy-Weinberg equilibrium (\code{"hw"}),
or a uniform distribution (\code{"uniform"}).}

\item{...}{Additional arguments to pass to \code{\link{updog_vanilla}}.}
}
\description{
This function fits a hierarchical model to sequence counts from
collection of siblings --- or a population of individuals
in Hardy-Weinberg equilibrium --- and returns genotyped information. The
hierarchy comes from either the fact that they share the same parents or they come
from a population in Hardy-Weinberg equilibrium. If
you also have parental sequencing data, then you can include this
to improve estimates. This is the simple version of \code{updog} with only a limited number of parameters.
To see the full list of options, see \code{\link{updog_vanilla}}.
}
\seealso{
\code{\link{updog_vanilla}} for more parameter options.
}
\author{
David Gerard
}
