% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_updog_ideas.R
\name{updog_iterate}
\alias{updog_iterate}
\title{Iterate between indiividual level estimates and joint averaging.}
\usage{
updog_iterate(ocounts, osize, qarray, r1vec, r2vec, seq_error = 0.01,
  itermax = 1000, tol = 10^-4)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{qarray}{An array of proportions. Each dimension size is equal to the ploidy plus 1.}

\item{r1vec}{The posterior probabilities of the genotypes of parent 1 conditional only
on the sequence data from parent 1.}

\item{r2vec}{The posterior probabilities of the genotypes of parent 2 conditional only
on the sequence data from parent 2.}

\item{seq_error}{A non-negative numeric. This is the known
sequencing error rate. The default is to estimate this
using data that are all approximately the reference allele.}

\item{itermax}{The number of times to iterate between the two approaches.}

\item{tol}{A positive numeric. The tolderance for the stopping criteria.}
}
\description{
Iterate between indiividual level estimates and joint averaging.
}
\author{
David Gerard
}
