% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog.R
\name{updog_maximize}
\alias{updog_maximize}
\title{Estimate parental genotypes using offspring genotypes.}
\usage{
updog_maximize(ocounts, osize, qarray, r1vec, r2vec, pk, pival = 0.99,
  alpha = 0.1, beta = 0.1, est_fudge = TRUE, tol = 10^-4,
  itermax = 1000, update_geno = FALSE, update_pi = FALSE,
  update_beta = FALSE)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{qarray}{An three-way array of proportions. The (i, j, k)th
element is the probability of an offspring having k - 1
reference alleles given that parent 1 has i - 1 refrerence
alleles and parent 2 has j - 1 reference alleles. Each
dimension of the array is \code{ploidy + 1}.}

\item{r1vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 1 has i -1 copies of allele
A.}

\item{r2vec}{A vector of prior probabilities whose ith element is
the prior probability that parent 2 has i -1 copies of allele
A.}

\item{pk}{The parameter space. Should all be between 0 and 1.}

\item{pival}{A fudge factor. The probability of not begin a
mistake. Estimated if \code{est_fudge = TRUE}. Set to 0.99. Only used if \code{est_fudge = TRUE}.}

\item{alpha}{The alpha parameter in the beta distribution of the outliers. Only used if \code{est_fudge = TRUE}.}

\item{beta}{The beta parameter in the beta distribution of the outliers. Only used if \code{est_fudge = TRUE}.}

\item{est_fudge}{A logical. Should we estimate the fudge factor
\code{TRUE} or not \code{FALSE}?}

\item{tol}{A positive numeric. The tolerance for the stopping criterion.}

\item{itermax}{A positive integer. The maximum number of iterations to run in the EM.}

\item{update_geno}{A logical. Update the parental genotypes?}

\item{update_pi}{A logical. Update the mixing proporiton?}

\item{update_beta}{A logical. Update the beta distribution?}
}
\description{
Estimate parental genotypes using offspring genotypes.
}
\author{
David Gerard
}
