% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{updog_vanilla}
\alias{updog_vanilla}
\title{Using Parental Data for Offspring Genotyping.}
\usage{
updog_vanilla(ocounts, osize, ploidy, p1counts = NULL, p1size = NULL,
  p2counts = NULL, p2size = NULL, commit_num = 4, min_disp = 0,
  print_val = FALSE, update_outmean = FALSE, update_outdisp = FALSE,
  update_outprop = TRUE, update_bias_val = TRUE, update_seq_error = TRUE,
  update_od_param = TRUE, update_pgeno = TRUE, p1geno = 3, p2geno = 3,
  seq_error = 0.01, od_param = 0.01, bias_val = 1, out_prop = 0.001,
  out_mean = 0.5, out_disp = 1/3, non_mono_max = 2, bound_bias = FALSE,
  seq_error_mean = -4.7, seq_error_sd = 1, bias_val_mean = 0,
  bias_val_sd = 0.7, allele_freq = 0.5, model = c("f1", "s1", "hw",
  "uniform"), verbose = FALSE)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{ploidy}{A positive integer. The number of copies of the
genome in the species. This is the assumed to be the same
for all individuals.}

\item{p1counts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith
sample of parent 1.  If \code{NULL} then the prior
probabilities on parent 1's genotype will default to uniform.}

\item{p1size}{A vector of positive integers. The ith element is the
total number of reads in the ith sample of parent 1.  If
\code{NULL} then the prior probabilities on parent 1's genotype
will default to uniform.}

\item{p2counts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith
sample of parent 2.  If \code{NULL} then the prior
probabilities on parent 2's genotype will default to uniform.}

\item{p2size}{A vector of positive integers. The ith element is the
total number of reads in the ith sample of parent 2.  If
\code{NULL} then the prior probabilities on parent 2's genotype
will default to uniform.}

\item{commit_num}{The number of consecutive iterations where the parental
genotypes do not change before we commit to those parental genotypes.}

\item{min_disp}{The minimum value for the over-dispersion parameter of the
outlier distribution. If this gets too small, then we can be overly confident in
points being outliers.}

\item{print_val}{A logical. Should we print the updates (\code{TRUE}) or not (\code{FALSE})?}

\item{update_outmean}{A logical. Should we update \code{out_mean}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_outdisp}{A logical. Should we update \code{out_mean}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_outprop}{A logical. Should we update \code{out_prop}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_bias_val}{A logical. Should we update \code{bias_val}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_seq_error}{A logical. Should we update \code{seq_error}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_od_param}{A logical. Should we update \code{od_param}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_pgeno}{A logical. Should we update \code{p1geno} and \code{p1geno}
(\code{TRUE}) or not (\code{FALSE})?}

\item{p1geno}{The initial value of the first parental genotype.}

\item{p2geno}{The initial value of the second parental genotype.}

\item{seq_error}{A non-negative numeric. This is the known
sequencing error rate. The default is to estimate this
using data that are all approximately the reference allele.}

\item{od_param}{The initial value of the overdispersion parameter.}

\item{bias_val}{The initial value of the bias parameter.}

\item{out_prop}{The initial value of the proportion of points that are outliers.}

\item{out_mean}{The initial value of the mean of the outlier distribution.}

\item{out_disp}{The initial value of the over-dispersion parameter of the outlier distribution.}

\item{non_mono_max}{The maximum number of iterations to allow non-monotonicity of likelihood.}

\item{bound_bias}{A logical. Should we bound \code{bias_val} by a somewhat arbitrary value
(\code{TRUE}) or not (\code{FALSE})?}

\item{seq_error_mean}{The mean of the logit-normal prior on the sequencing error rate, which corresponds
to \code{parvec[2]}. Set \code{seq_error_sd = Inf} to have no penalty on the sequencing error rate.
The default is -4.7, which roughly corresponds to a mean sequencing error value of 0.009. If you want
to constain \code{seq_error} to be zero, you need to set \code{update_seq_error = FALSE},
\code{seq_error = 0}, and\code{seq_error_mean = -Inf}.}

\item{seq_error_sd}{The standard deviation of the logit-normal prior on the sequencing error rate, which
corresponds to \code{parvec[2]}. The default is 1, which at three standard deviations is about a sequencing
error rate of 0.15. Set \code{seq_error_sd = Inf} to have no penalty on the sequencing error rate.}

\item{bias_val_mean}{The prior mean on the log of \code{bias_val} (corresponding to \code{parvec[1]}).
Set \code{bias_val_sd = Inf} to have no penalty on the bias parameter.}

\item{bias_val_sd}{The prior standard deviation on the log of \code{bias_val}
(corresponding to \code{parvec[1]}).
Set \code{bias_val_sd = Inf} to have no penalty on the bias parameter.}

\item{allele_freq}{The allele frequency if \code{model = "hw"}.}

\item{model}{The model for the genotype distribution. Do we assume an
F1 population (\code{"f1"}), an S1 population (\code{"s1"}), Hardy-Weinberg equilibrium (\code{"hw"}),
or a uniform distribution (\code{"uniform"}).}

\item{verbose}{A logical. Should we output a lot more (\code{TRUE}) or not (\code{FALSE})?}
}
\value{
A list of class \code{updog} with some or all of the following elements:
    \describe{
        \item{\code{ogeno}}{A vector. Each element of which is the maximum a posteriori estimate of each individual's genotype.}
        \item{\code{maxpostprob}}{A vector. The maximum posterior probability of a genotype for each individual.}
        \item{\code{postmean}}{A vector. The posterior mean genotype for each individual.}
        \item{\code{bias_val}}{The estimated bias parameter. This is a value greater than 0 which is the ratio of the probability of correctly mapping a read containing the alternative allele to the probability of correctly mapping a read containing the reference allele. A value of 1 indicates no bias. A value less than one indicates bias towards the reference alelle. A value greater than 1 indiciates bias towards the alternative allele.}
        \item{\code{seq_error}}{The estimated sequencing error rate. This is between 0 and 1.}
        \item{\code{od_param}}{The estimated overdispersion parameter. Also known as the "intra-class correlation", this is the overdispersion parameter in the underlying beta of the beta-binomial distribution of the counts. Between 0 and 1, a value closer to 0 indicates less overdispersion and a value greater than 1 indicates greater overdispersion. In real data, we we typically see estimates between 0 and 0.01.}
        \item{\code{p1geno}}{The estimated genotype of one parent. The number of copies of the reference allele one of the parents has.}
        \item{\code{p1geno}}{The estimated genotype of the other parent. The number of copies of the reference allele the other parent has.}
        \item{\code{allele_freq}}{The estimated allele-frequency of the reference allele. This is the binomial proportion. Between 0 and 1, a value closer to 1 indicates a larger amount of reference alleles in the population.}
        \item{\code{out_prop}}{The estimated proportion of points that are outliers.}
        \item{\code{out_mean}}{The estimated mean of the outlier distribution. The outlier distribution is beta-binomial.}
        \item{\code{out_disp}}{The estimated overdispersion parameter of the outlier distribution. This is the "intra-class correlation" parameter of the beta-binomial outlier distribution.}
        \item{\code{prob_out}}{A vector. Each element of which is the posterior probability that a point is an outlier.}
        \item{\code{prob_ok}}{The posterior probability that a point is a non-outlier.}
        \item{\code{p1_prob_out}}{The posterior probability that parent 1 is an outlier.}
        \item{\code{p2_prob_out}}{The posterior probability that parent 2 is an outlier.}
        \item{\code{num_iter}}{The number of iterations the optimization program was run.}
        \item{\code{convergence}}{\code{1} if we reached \code{maxiter} and \code{0} otherwise.}
        \item{\code{llike}}{The final log-likelihood of the estimates.}
        \item{\code{hessian}}{The negative-Fisher information under the parameterization (s, ell, r), where s = log(bias_val) = log(d), ell = logit(seq_error) = logit(eps), and r = - logit(od_param) = - logit(tau). If you want standard errors for these parameters (in the described parameterization), simply take the negative inverse of the hessian.}
        \item{\code{input}}{A list with the input counts (\code{ocounts}), the input sizes (\code{osize}), input parental counts (\code{p1counts} and \code{p2counts}), input parental sizes (\code{p2size} and \code{p1size}), the ploidy (\code{ploidy}) and the model (\code{model}).}
        \item{\code{log_bias}}{The log of \code{bias_val}}
        \item{\code{logit_seq_error}}{The logit of \code{seq_error}. I.e. \code{log(seq_error / (1 - seq_error))}}
        \item{\code{neg_logit_od_param}}{The negative logit of \code{od_param}. I.e. \code{log((1 - od_param) / od_param)}}
        \item{\code{covmat}}{The observed Fisher information matrix of \code{c(log_bias, logit_seq_error, neg_logit_od_param)}. This can be used as the covariance matrix of these estimates.}
    }
}
\description{
This function fits a hierarchical model to sequence counts from a
collection of siblings --- or a population of individuals
in Hardy-Weinberg equilibrium --- and returns genotyped information. The
hierarchy comes from either the fact that they share the same parents or they come
from a population in Hardy-Weinberg equilibrium. If
you also have parental sequencing data, then you can include this
to improve estimates.
}
\details{
The key improvements in \code{updog} are its abilities to account for common features in
GBS data: sequencing error rate, read-mapping bias, overdispersion, and outlying points.
}
\seealso{
\code{\link{plot.updog}} For plotting the output of \code{\link{updog_vanilla}}.
}
\author{
David Gerard
}
