% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_algorithm.R
\name{get_q_array}
\alias{get_q_array}
\title{Derive the probabilities of an offspring's genotype given its parental
genotypes for all possible combinations of parental and offspring genotypes.}
\usage{
get_q_array(ploidy)
}
\arguments{
\item{ploidy}{A positive integer. The ploidy of the species.}
}
\description{
Derive the probabilities of an offspring's genotype given its parental
genotypes for all possible combinations of parental and offspring genotypes.
}
\author{
David Gerard
}

