% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_offspring_original}
\alias{grad_offspring_original}
\title{Gradient of \code{\link{obj_offspring}} using the original parameterization.}
\usage{
grad_offspring_original(ocounts, osize, ploidy, p1geno, p2geno, d, eps, tau)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{p1geno}{The first parental genotype. The number of copies of the
reference allele the first parent has.}

\item{p2geno}{The second parental genotype. The number of copies of the
reference allele the second parent has.}

\item{d}{The sequencing bias parameter.}

\item{eps}{The sequencing error rate}

\item{tau}{The overdisperion parameter.}
}
\description{
Gradient of \code{\link{obj_offspring}} using the original parameterization.
}
\author{
David Gerard
}
