% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_offspring_weights}
\alias{grad_offspring_weights}
\title{Gradient of \code{\link{obj_offspring_weights_reparam}}}
\usage{
grad_offspring_weights(ocounts, osize, weight_vec, ploidy, p1geno, p2geno, d,
  ell, h)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{weight_vec}{A vector of weights between 0 and 1 (do not need to add up to 1).}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{p1geno}{The first parental genotype. The number of copies of the
reference allele the first parent has.}

\item{p2geno}{The second parental genotype. The number of copies of the
reference allele the second parent has.}

\item{d}{The bias term}

\item{ell}{The logit of the sequencing error rate}

\item{h}{We have h = (1 - tau) / tau}
}
\description{
Gradient of \code{\link{obj_offspring_weights_reparam}}
}
