% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_offspring_weights_reparam}
\alias{obj_offspring_weights_reparam}
\title{Reparameterization of \code{\link{obj_offspring_weights}}.}
\usage{
obj_offspring_weights_reparam(ocounts, osize, weight_vec, ploidy, p1geno,
  p2geno, s, ell, r)
}
\arguments{
\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{weight_vec}{A vector of numerics between 0 and 1. They don't have to sum to 1.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{p1geno}{The first parental genotype. The number of copies of the
reference allele the first parent has.}

\item{p2geno}{The second parental genotype. The number of copies of the
reference allele the second parent has.}

\item{s}{Same as \code{exp(bias_val)} in \code{\link{obj_offspring}}.}

\item{ell}{We have \code{seq_error = expit(ell)} from \code{\link{obj_offspring}}.}

\item{r}{Same as \code{log((1.0 - od_param) / od_param)} from \code{\link{obj_offspring}}.}
}
\description{
Reparameterization of \code{\link{obj_offspring_weights}}.
}
\author{
David Gerard
}
