% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.updog}
\alias{plot.updog}
\title{Draw a genotype plot from the output of \code{\link{updog}}.}
\usage{
\method{plot}{updog}(x, gg = requireNamespace("ggplot2", quietly = TRUE),
  plot_beta = FALSE, ask = TRUE, use_colorblind = FALSE,
  show_maxpostprob = FALSE, show_ogeno = TRUE, show_outlier = TRUE, ...)
}
\arguments{
\item{x}{The output from \code{\link{updog}}.}

\item{gg}{Should we use ggplot2 to plot the genotypes
(\code{TRUE}), or not (\code{FALSE}).  If ggplot2 is present, then
this defaults to \code{TRUE}. If it is not present, then it
defaults to \code{FALSE}.}

\item{plot_beta}{A logical. If true, then we'll also plot the
estimated beta density of the outlier model, followed by the
estimated beta distributions of the overdispersion models.}

\item{ask}{A logical. Should we ask before continuing on to the
next plot (\code{TRUE}) or not (\code{FALSE})?}

\item{use_colorblind}{A logical. Should we use a colorblind safe palette (\code{TRUE}),
or not (\code{FALSE})?}

\item{show_maxpostprob}{A logical. Should we scale the sizes by \code{x$maxpostprob} (\code{TRUE}),
or not (\code{FALSE})?}

\item{show_ogeno}{A logical. Should we color code by \code{x$ogeno} (\code{TRUE}),
or not (\code{FALSE})?}

\item{show_outlier}{A logical. Should we scale the transparency by \code{x$prob_ok} (\code{TRUE}),
or not (\code{FALSE})?}

\item{...}{Not used.}
}
\value{
A plot object.
}
\description{
If ggplot2 is installed, then this function will use it to make
the genotype plot.  Otherwise, "classic" R base graphics will be
used. The ggplot2 version is made using the function
\code{\link{plot_geno}}. The "classic" R base graphics version is made
using the function \code{\link{plot_geno_base}}.
}
\details{
The first plot is what we call a "genotype plot". On the x-axis is the
number of "a" reads and on the y-axis is the number of "A" reads, where
"a" and "A" are the possible alleles. If available, the observations are
colorcoded by estimated genotype (via the maximum a posteriori
classifier), their transparency is proportional to the the posterior
probability that a point is an outlier, and their size is (possibly) scaled
by their maximum posterior probability. The lines are defined by the
equation
\deqn{A / (A + a) = p,}
where \eqn{A} is the number of "A" reads, \eqn{a} is the number of "a" reads,
and \eqn{p} is the proportion of counts we would expect to observe on average
given the genotype of an individual, the sequencing error rate, and the
read-mapping bias.

If there is parental data, these are also plotted with crosses and plusses.
If the parents have larger counts than the offspring, then (at least
when \code{gg = TRUE}) the parental
counts are scaled until they just reach the plot. The scaled parental counts
are labeled on the plot with a semi-transparent "(scaled)".

The second plot is the distribution of outlying points.

The third plot contains the underlying beta densities for the different
genotypes.
}
\author{
David Gerard
}
