% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog.R
\name{plot_geno}
\alias{plot_geno}
\title{Make those awesome plots that Felipe showed me, but using ggplot2.}
\usage{
plot_geno(ocounts, osize, ploidy, p1counts = NULL, p1size = NULL,
  p2counts = NULL, p2size = NULL, col = NULL, seq_error = 0.01,
  prob_ok = NULL, maxpostprob = NULL)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The number of
reference alleles observed in the offspring.}

\item{osize}{A vector of positive integers. The total number of
reads in the offspring.}

\item{ploidy}{A non-negative integer. The ploidy of the species.}

\item{p1counts}{A vector of non-negative integers. The number of
reference alleles observed in parent 1.}

\item{p1size}{A vector of positive integers. The total number of
reads in parent 1.}

\item{p2counts}{A vector of non-negative integers. The number of
reference alleles observed in parent 2.}

\item{p2size}{A vector of positive integers. The total number of
reads in parent 2.}

\item{col}{The color labels.}

\item{seq_error}{The average sequencing error rate.}

\item{prob_ok}{A vector of posterior probabilities of not being a mistake.}

\item{maxpostprob}{A vector of the posterior probabilities of begin at the modal probability.}
}
\description{
Make those awesome plots that Felipe showed me, but using ggplot2.
}
\author{
David Gerard
}

