% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.updog}
\alias{summary.updog}
\title{Summary method for class "\code{updog}".}
\usage{
\method{summary}{updog}(object, ...)
}
\arguments{
\item{object}{An \code{updog} object. Usually what has been returned from \code{\link{updog}}.}

\item{...}{Not used.}
}
\value{
A list with three elements:

    \code{prop_ok}: The estimated proportion of observations that are outliers.

    \code{genotypes}: Counts for how many of each estimated genotype are observed.

    \code{summ_prob}: A data frame with summaries on two variables: The maximum posterior
    probability of a genotype that each observation has (\code{maxpostprob}), and the posterior probability
    of not being an outlier (\code{prob_ok})
}
\description{
Summary method for class "\code{updog}".
}
\author{
David Gerard
}

