% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{update_good}
\alias{update_good}
\title{Update the OK points}
\usage{
update_good(parvec, ocounts, osize, weight_vec, ploidy, p1geno = NULL,
  p2geno = NULL, bound = TRUE)
}
\arguments{
\item{parvec}{A vector of three elements, s, ell, and r. We have s = log(bias_val) = log(d),
ell = logit(seq_error) = logit(eps), and r = 1 / logit(od_param) = 1 / logit(tau).}

\item{ocounts}{The observed counts of the refernce
allele for each individual.}

\item{osize}{The observed number of reads for each
individuals.}

\item{weight_vec}{A vector of weights obtained via the E-step.}

\item{ploidy}{An integer. The ploidy of the species. This is assumed
to be the same for each individual.}

\item{p1geno}{The first parental genotype. The number of copies of the
reference allele the first parent has.}

\item{p2geno}{The second parental genotype. The number of copies of the
reference allele the second parent has.}

\item{bound}{A logical. Should we bound the bias parameter by a somewhat arbitrary value
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
Update the OK points
}
\author{
David Gerard
}
