% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_algorithm.R
\name{updog}
\alias{updog}
\title{Using Parental DNA for Offspring Genotyping.}
\usage{
updog(ocounts, osize, p1counts, p1size, p2counts, p2size, ploidy)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element is the
number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the total
number of reads for the ith child.}

\item{p1counts}{A vector of non-negative integers. The ith element is the
number of reads of the reference allele in the ith sample of parent 1.}

\item{p1size}{A vector of positive integers. The ith element is the total
number of reads in the ith sample of parent 1.}

\item{p2counts}{A vector of non-negative integers. The ith element is the
number of reads of the reference allele in the ith sample of parent 2.}

\item{p2size}{A vector of positive integers. The ith element is the total
number of reads in the ith sample of parent 2.}

\item{ploidy}{A positive integer. The number of copies of the genome in the
species.}
}
\description{
This function takes parental and offspring sequence counts and returns
posterior probabilities on the genotype of all offspring. This isn't fully
Bayesian because I only update the parental probabilities with just their
sequencing data.
}
\details{
What is updog? Nothing much, what's up with you dawg!
}
\author{
David Gerard
}

