% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog.R
\name{updog}
\alias{updog}
\title{Using Parental Data for Offspring Genotyping.}
\usage{
updog(ocounts, osize, ploidy, model = c("f1", "s1", "hw", "uniform"), ...)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{ploidy}{A positive integer. The number of copies of the
genome in the species.}

\item{model}{The model for the genotype distribution. Do we assume an
F1 population (\code{"f1"}), an S1 population (\code{"s1"}), Hardy-Weinberg equilibrium (\code{"hw"}),
or a uniform distribution (\code{"uniform"}).}

\item{...}{Additional arguments to pass to \code{\link{updog_vanilla}}.}
}
\value{
A list of class \code{updog} with some or all of the following elements:
    \itemize{
        \item{\code{bias_val}:}{ The estimated bias parameter.}
        \item{\code{seq_error}:}{ The estimated sequencing error rate.}
        \item{\code{od_param}:}{ The estimated overdispersion parameter.}
        \item{\code{p1geno}:}{ The estimated genotype of one parent.}
        \item{\code{p1geno}:}{ The estimated genotype of the other parent.}
        \item{\code{out_prop}:}{ The estimated proportion of points that are outliers.}
        \item{\code{out_mean}:}{ The estimated mean of the outlier distribution.}
        \item{\code{out_disp}:}{ The estimated overdispersion parameter of the outlier distribution.}
        \item{\code{prob_out}:}{ A vector. Each element of which is the posterior probability that a point is an outlier.}
        \item{\code{allele_freq}:}{ The estimated allele-frequency of the reference allele.}
        \item{\code{p1_prob_out}:}{ The posterior probability that parent 1 is an outlier.}
        \item{\code{p2_prob_out}:}{ The posterior probability that parent 2 is an outlier.}
        \item{\code{num_iter}:}{ The number of iterations the optimization program was run.}
        \item{\code{convergence}:}{ 1 is we reached \code{maxiter} and 0 otherwise.}
        \item{\code{llike}:}{ The final log-likelihood of the estimates.}
        \item{\code{hessian}:}{ The Fisher information under the parameterization (s, ell, r), where s = log(bias_val) = log(d),
ell = logit(seq_error) = logit(eps), and r = - logit(od_param) = - logit(tau).}
        \item{\code{ogeno}:}{ A vector. Each element of which is the maximum a posteriori estimate of each individual genotype.}
        \item{\code{maxpostprob}:} {The maximum posterior probability of a genotype for each individual.}
        \item{\code{postmean}:}{ The posterior mean genotype for each individual.}
        \item{\code{prob_ok}:}{ The posterior probability that a point is a non-outlier.}
        \item{\code{input}:}{ A list with the input counts (\code{ocounts}), the input sizes (\code{osize}), input parental counts (\code{p1counts} and \code{p2counts}), input parental sizes (\code{p2size} and \code{p1size}), the ploidy (\code{ploidy}) and the model (\code{model}).}
    }
}
\description{
This function fits a hierarchical model to sequence counts from
collection of siblings --- or a population of individuals
in Hardy-Weinberg equilibrium --- and returns genotyped information. The
hierarchy comes from either the fact that they share the same parents or they come
from a population in Hardy-Weinberg equilibrium. If
you also have parental sequencing data, then you can include this
to improve estimates. This is the simple version of \code{updog} with only a limited number of parameters.
To see the full list of options, see \code{\link{updog_vanilla}}.
}
\details{
The key improvements in \code{updog} are its abilities to account for common features in
GBS data: sequencing error rate, read-mapping bias, overdispersion, and outlying points.
}
\seealso{
\code{\link{updog_vanilla}} for more parameter options.
}
\author{
David Gerard
}
