% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{updog_update_all}
\alias{updog_update_all}
\title{This is the main optimization function for updog.}
\usage{
updog_update_all(ocounts, osize, ploidy, p1counts = NULL, p1size = NULL,
  p2counts = NULL, p2size = NULL, print_val = TRUE, tol = 10^-4,
  maxiter = 500, commit_num = 4, min_disp = 0, update_outmean = FALSE,
  update_outdisp = FALSE, update_outprop = TRUE, update_bias_val = TRUE,
  update_seq_error = TRUE, update_od_param = TRUE, update_pgeno = TRUE,
  p1geno = round(ploidy/2), p2geno = round(ploidy/2), seq_error = 0.01,
  od_param = 0.01, bias_val = 1, out_prop = 0.001, out_mean = 0.5,
  out_disp = 1/3, non_mono_max = 2, bound_bias = FALSE,
  seq_error_mean = -4.7, seq_error_sd = 1, bias_val_mean = 0,
  bias_val_sd = 0.7, allele_freq = 0.5, model = c("f1", "s1", "hw",
  "uniform"))
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{ploidy}{A positive integer. The number of copies of the
genome in the species. This is the assumed to be the same
for all individuals.}

\item{p1counts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith
sample of parent 1.  If \code{NULL} then the prior
probabilities on parent 1's genotype will default to uniform.}

\item{p1size}{A vector of positive integers. The ith element is the
total number of reads in the ith sample of parent 1.  If
\code{NULL} then the prior probabilities on parent 1's genotype
will default to uniform.}

\item{p2counts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele in the ith
sample of parent 2.  If \code{NULL} then the prior
probabilities on parent 2's genotype will default to uniform.}

\item{p2size}{A vector of positive integers. The ith element is the
total number of reads in the ith sample of parent 2.  If
\code{NULL} then the prior probabilities on parent 2's genotype
will default to uniform.}

\item{print_val}{A logical. Should we print the updates (\code{TRUE}) or not (\code{FALSE})?}

\item{tol}{The stopping criterion}

\item{maxiter}{The maximum number of iterations}

\item{commit_num}{The number of consecutive iterations where the parental
genotypes do not change before we commit to those parental genotypes.}

\item{min_disp}{The minimum value for the over-dispersion parameter of the
outlier distribution. If this gets too small, then we can be overly confident in
points being outliers.}

\item{update_outmean}{A logical. Should we update \code{out_mean}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_outdisp}{A logical. Should we update \code{out_mean}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_outprop}{A logical. Should we update \code{out_prop}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_bias_val}{A logical. Should we update \code{bias_val}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_seq_error}{A logical. Should we update \code{seq_error}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_od_param}{A logical. Should we update \code{od_param}
(\code{TRUE}) or not (\code{FALSE})?}

\item{update_pgeno}{A logical. Should we update \code{p1geno} and \code{p1geno}
(\code{TRUE}) or not (\code{FALSE})?}

\item{p1geno}{The initial value of the first parental genotype.}

\item{p2geno}{The initial value of the second parental genotype.}

\item{seq_error}{The initial value of the sequencing error rate.}

\item{od_param}{The initial value of the overdispersion parameter.}

\item{bias_val}{The initial value of the bias parameter.}

\item{out_prop}{The initial value of the proportion of points that are outliers.}

\item{out_mean}{The initial value of the mean of the outlier distribution.}

\item{out_disp}{The initial value of the over-dispersion parameter of the outlier distribution.}

\item{non_mono_max}{The maximum number of iterations to allow non-monotonicity of likelihood.}

\item{bound_bias}{A logical. Should we bound \code{bias_val} by a somewhat arbitrary value
(\code{TRUE}) or not (\code{FALSE})?}

\item{seq_error_mean}{The mean of the logit-normal prior on the sequencing error rate, which corresponds
to \code{parvec[2]}. Set \code{seq_error_sd = Inf} to have no penalty on the sequencing error rate.
The default is -4.7, which roughly corresponds to a mean sequencing error value of 0.009. If you want
to constain \code{seq_error} to be zero, you need to set \code{update_seq_error = FALSE},
\code{seq_error = 0}, and\code{seq_error_mean = -Inf}.}

\item{seq_error_sd}{The standard deviation of the logit-normal prior on the sequencing error rate, which
corresponds to \code{parvec[2]}. The default is 1, which at three standard deviations is about a sequencing
error rate of 0.15. Set \code{seq_error_sd = Inf} to have no penalty on the sequencing error rate.}

\item{bias_val_mean}{The prior mean on the log of \code{bias_val} (corresponding to \code{parvec[1]}).
Set \code{bias_val_sd = Inf} to have no penalty on the bias parameter.}

\item{bias_val_sd}{The prior standard deviation on the log of \code{bias_val}
(corresponding to \code{parvec[1]}).
Set \code{bias_val_sd = Inf} to have no penalty on the bias parameter.}

\item{allele_freq}{The allele frequency if \code{model = "hw"}.}

\item{model}{The model for the genotype distribution. Do we assume an
F1 population (\code{"f1"}), an S1 population (\code{"s1"}), Hardy-Weinberg equilibrium (\code{"hw"}),
or a uniform distribution (\code{"uniform"}).}
}
\value{
A list of the following elements
    \itemize{
        \item{\code{bias_val}:}{ The estimated bias parameter.}
        \item{\code{seq_error}:}{ The estimated sequencing error rate.}
        \item{\code{od_param}:}{ The estimated overdispersion parameter.}
        \item{\code{p1geno}:}{ The estimated genotype of one parent.}
        \item{\code{p1geno}:}{ The estimated genotype of the other parent.}
        \item{\code{out_prop}:}{ The estimated proportion of points that are outliers.}
        \item{\code{out_mean}:}{ The estimated mean of the outlier distribution.}
        \item{\code{out_disp}:}{ The estimated overdispersion parameter of the outlier distribution.}
        \item{\code{prob_out}:}{ A vector. Each element of which is the posterior probability that a point is an outlier.}
        \item{\code{allele_freq}:}{ The estimated allele-frequency of the reference allele.}
        \item{\code{p1_prob_out}:}{ The posterior probability that parent 1 is an outlier.}
        \item{\code{p2_prob_out}:}{ The posterior probability that parent 2 is an outlier.}
        \item{\code{num_iter}:}{ The number of iterations the optimization program was run.}
        \item{\code{convergence}:}{ 1 is we reached \code{maxiter} and 0 otherwise.}
        \item{\code{llike}:}{ The final log-likelihood of the estimates.}
        \item{\code{hessian}:}{ The Fisher information under the parameterization (s, ell, r), where s = log(bias_val) = log(d),
ell = logit(seq_error) = logit(eps), and r = - logit(od_param) =  - logit(tau).}
    }
}
\description{
This is the main optimization function for updog.
}
\author{
David Gerard
}
