% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{updog_vanilla}
\alias{updog_vanilla}
\title{This is a vanilla version of updog that will optimize everything, assuming a uniform prior on the parental
genotypes and not assuming that we have parental sequence data, and then return the posterior summaries.}
\usage{
updog_vanilla(ocounts, osize, ploidy, print_val)
}
\arguments{
\item{ocounts}{A vector of non-negative integers. The ith element
is the number of reads of the reference allele the ith child.}

\item{osize}{A vector of positive integers. The ith element is the
total number of reads for the ith child.}

\item{ploidy}{A positive integer. The number of copies of the
genome in the species.}

\item{print_val}{Should we print the updates?}
}
\description{
This is a vanilla version of updog that will optimize everything, assuming a uniform prior on the parental
genotypes and not assuming that we have parental sequence data, and then return the posterior summaries.
}
\author{
David Gerard
}
